Tog bort main activity då den var onödig. Parkodat med Moa.
