+++ app/src/main/java/com/example/gameboy/GameActivity.kt
@@ -1,9 +1,11 @@
 package com.example.gameboy
 
+import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.fragment.app.commit
 import com.example.gameboy.databinding.ActivityGameBinding
+import com.google.firebase.auth.FirebaseAuth
 
 class GameActivity : AppCompatActivity(), PongFragment.GameListener,
     GameSelectFragment.GameListener, Ball.GameListener {
@@ -19,6 +21,12 @@ class GameActivity : AppCompatActivity(), PongFragment.GameListener,
         supportFragmentManager.commit {
             replace(R.id.frame3, GameSelectFragment())
         }
+
+        binding.logoutButton.setOnClickListener {
+            logoutUser()
+
+            // Optionally, navigate to the login screen or perform other actions
+        }
     }
 
     override fun startPong() {
@@ -31,6 +39,13 @@ class GameActivity : AppCompatActivity(), PongFragment.GameListener,
         supportFragmentManager.commit {
             replace(R.id.frame3, PongFragment())
         }
-    }
 
-}
+    }
+        fun logoutUser() {
+            val auth = FirebaseAuth.getInstance()
+            auth.signOut()
+            val intent = Intent(this, SignInActivity::class.java)
+            startActivity(intent)
+            finish()
+        }
+    }
