+++ app/src/main/res/layout/activity_game.xml
@@ -6,17 +6,17 @@
     android:layout_height="match_parent"
     tools:context=".GameActivity">
 
-    <ImageView
-        android:id="@+id/imageView"
-        android:layout_width="1840dp"
-        android:layout_height="735dp"
-        android:src="@drawable/nintendo_switch7"
+    <Button
+        android:id="@+id/logoutButton"
+        android:text="Log out"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.498"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-
-        />
+        app:layout_constraintVertical_bias="0.055" />
 
     <FrameLayout
         android:id="@+id/frame3"
